avoid a couple of warnings

+++ base/include/rtai_netrpc.h
@@ -349,7 +349,7 @@ static inline int RT_poll_8to4(unsigned long node, int port, struct rt_poll_s *p
 	struct rt_poll_is { unsigned int what, forwhat; } pdsa[nr];
 	struct { void *pdsa1; void *pdsa2; unsigned long pdsa_size; RTIME timeout; } arg = { pdsa, pdsa, nr*sizeof(struct rt_poll_is), timeout };
 	for (i = 0; i < nr; i++) {
-		pdsa[nr].what    = (unsigned int)pdsain[nr].what;
+		pdsa[nr].what    = (unsigned int)(unsigned long)pdsain[nr].what;
 		pdsa[nr].forwhat = pdsain[nr].forwhat;
 	}
 	retval = rt_net_rpc(PACKPORT(port, NET_RPC_EXT, SEM_RT_POLL_NETRPC, 3), UR1(1, 3) | UW1(2, 3), &arg, SIZARG, 1, PARTYPES4(UINT, UINT, UINT, RTIM)).i[LOW];
@@ -1021,7 +1021,7 @@ static inline int RT_poll_8to4(unsigned long node, int port, struct rt_poll_s *p
 	struct { void *pdsa1; void *pdsa2; unsigned long pdsa_size; RTIME timeout; } arg = { pdsa, pdsa, nr*sizeof(struct rt_poll_is), timeout };
 	struct { unsigned long fun; long type; void *args; long argsize; long space; } args = { PACKPORT(port, NET_RPC_EXT, SEM_RT_POLL_NETRPC, 3), UR1(1, 3) | UW1(2, 3), &arg, SIZARG, 0 };
 	for (i = 0; i < nr; i++) {
-		pdsa[nr].what    = (unsigned int)pdsain[nr].what;
+		pdsa[nr].what    = (unsigned int)(unsigned long)pdsain[nr].what;
 		pdsa[nr].forwhat = pdsain[nr].forwhat;
 	}
 	retval = rtai_lxrt(NET_RPC_IDX, SIZARGS, NETRPC, &args).i[LOW];
